fix: get program from storage
